requirements.txt
certifi==2022.12.7
chardet==5.1.0
docutils==0.19
filestack-python==3.5.0
idna==3.4

Pygments==2.14.0
requests==2.28.2
trafaret==2.1.1
urllib3==1.26.14
Kivy==2.1.0
Kivy-Garden==0.1.5